add support for multi-dimensional packing and unpack
