+++ docs/source/conf.py
@@ -188,7 +188,7 @@ from sphinx_gallery.sorting import FileNameSortKey
 sphinx_gallery_conf = {
         'examples_dirs': ['../../samples', '../../tutorials'],
         'gallery_dirs': ['samples', 'tutorials'],
-        'ignore_pattern': '(.*data\.py)|(.*merlinc\.py)|(example.*py)',
+        'ignore_pattern': '(.*data\.py)|(.*merlinc\.py)|(example.*py)|(test.*py)',
         'filename_pattern': '(digitrec\.py)|(tutorial.*py)',
         'within_subsection_order': FileNameSortKey,
 }
