+++ tests/test_api_schedule.py
@@ -1,0 +1,136 @@
+import heterocl as hcl
+import numpy as np
+
+"""
+Testing API: make schedule
+"""
+
+def test_base_update():
+    A = hcl.placeholder((10,))
+    B = hcl.placeholder((10,))
+
+    def algorithm(A, B):
+        U = hcl.update(B, lambda x: A[x] + 1)
+
+    s = hcl.create_schedule([A, B], algorithm)
+    f = hcl.build(s)
+
+    _A = hcl.asarray(np.random.randint(100, size=(10,)), dtype = hcl.Int(32))
+    _B = hcl.asarray(np.zeros(10), dtype = hcl.Int(32))
+
+    f(_A, _B)
+
+    _A = _A.asnumpy()
+    _B = _B.asnumpy()
+
+    for i in range(10):
+        assert(_B[i] == _A[i] + 1)
+
+def test_base_return():
+    A = hcl.placeholder((10,))
+
+    def algorithm(A):
+        return hcl.compute(A.shape, lambda x: A[x] + 1)
+
+    s = hcl.create_schedule([A], algorithm)
+    f = hcl.build(s)
+
+    _A = hcl.asarray(np.random.randint(100, size=(10,)), dtype = hcl.Int(32))
+    _B = hcl.asarray(np.zeros(10), dtype = hcl.Int(32))
+
+    f(_A, _B)
+
+    _A = _A.asnumpy()
+    _B = _B.asnumpy()
+
+    for i in range(10):
+        assert(_B[i] == _A[i] + 1)
+
+def test_base_return_multi():
+    A = hcl.placeholder((10,))
+
+    def algorithm(A):
+        B = hcl.compute(A.shape, lambda x: A[x] + 1)
+        C = hcl.compute(A.shape, lambda x: A[x] + 2)
+        return B, C
+
+    s = hcl.create_schedule([A], algorithm)
+    f = hcl.build(s)
+
+    _A = hcl.asarray(np.random.randint(100, size=(10,)), dtype = hcl.Int(32))
+    _B = hcl.asarray(np.zeros(10), dtype = hcl.Int(32))
+    _C = hcl.asarray(np.zeros(10), dtype = hcl.Int(32))
+
+    f(_A, _B, _C)
+
+    _A = _A.asnumpy()
+    _B = _B.asnumpy()
+    _C = _C.asnumpy()
+
+    for i in range(10):
+        assert(_B[i] == _A[i] + 1)
+        assert(_C[i] == _A[i] + 2)
+
+def test_compute():
+    A = hcl.placeholder((10,))
+    B = hcl.placeholder((10,))
+
+    def algorithm(A, B):
+        C = hcl.compute(A.shape, lambda x: A[x] + 1)
+        hcl.update(B, lambda x: C[x] + 1)
+
+    s = hcl.create_schedule([A, B], algorithm)
+    f = hcl.build(s)
+
+    _A = hcl.asarray(np.random.randint(100, size=(10,)), dtype = hcl.Int(32))
+    _B = hcl.asarray(np.zeros(10), dtype = hcl.Int(32))
+
+    f(_A, _B)
+
+    _A = _A.asnumpy()
+    _B = _B.asnumpy()
+
+    for i in range(10):
+        assert(_B[i] == _A[i] + 2)
+
+def test_schedule():
+    A = hcl.placeholder((10,))
+    B = hcl.placeholder((10,))
+
+    def algorithm(A, B):
+        hcl.update(B, lambda u0: A[u0] + 1, name="U")
+
+    s = hcl.create_schedule([A, B], algorithm)
+    s[algorithm.U].unroll(algorithm.U.axis[0])
+
+    s = hcl.lower(s)
+
+    assert 'unrolled "factor"=0 (u0, 0, 10)' in str(s)
+
+"""
+Testing API: make_scheme & make_schedule_from_scheme
+"""
+
+def test_resize():
+
+    def algorithm(A):
+        return hcl.compute(A.shape, lambda x: A[x] + 1, "B")
+
+    A = hcl.placeholder((10,), dtype = hcl.UInt(32))
+
+    scheme = hcl.create_scheme([A], algorithm)
+    scheme.downsize(algorithm.B, hcl.UInt(2))
+    s = hcl.create_schedule_from_scheme(scheme)
+    f = hcl.build(s)
+
+    a = np.random.randint(100, size=(10,))
+    _A = hcl.asarray(a, dtype = hcl.UInt(32))
+    _B = hcl.asarray(np.zeros(10), dtype = hcl.UInt(2))
+
+    f(_A, _B)
+
+    _A = _A.asnumpy()
+    _B = _B.asnumpy()
+
+    for i in range(10):
+        assert(_B[i] == (a[i] + 1)%4)
