+++ tutorials/test_tutorial.py
@@ -1,0 +1,4 @@
+def test_tutorial_01():
+
+    from tutorial_01_get_started import test
+    test()
